Move to BIS.Core
