+++ BIS.Core.Test/ArmaTextDeserializerTest.cs
@@ -1,6 +1,7 @@
+using BIS.Core.Serialization;
 using Xunit;
 
-namespace BIS.TXT.Test
+namespace BIS.Core.Test.Serialization
 {
     public class ArmaTextDeserializerTest
     {
