+++ bis-file-formats.sln
@@ -17,9 +17,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BIS.RTM", "BIS.RTM\BIS.RTM.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BIS.WRP", "BIS.WRP\BIS.WRP.csproj", "{2B43DF8A-8D46-45FE-AD10-63AF88E0B570}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BIS.TXT", "BIS.TXT\BIS.TXT.csproj", "{CFA8567B-FA5F-4FF5-A1CC-FAC5A73D9080}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BIS.TXT.Test", "BIS.TXT.Test\BIS.TXT.Test.csproj", "{1F36CE5E-843C-4960-8B74-B87DEAB0C730}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BIS.Core.Test", "BIS.Core.Test\BIS.Core.Test.csproj", "{6A91F1B2-5C8A-495A-824B-1BC7825B1939}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -55,14 +53,10 @@ Global
 		{2B43DF8A-8D46-45FE-AD10-63AF88E0B570}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2B43DF8A-8D46-45FE-AD10-63AF88E0B570}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2B43DF8A-8D46-45FE-AD10-63AF88E0B570}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CFA8567B-FA5F-4FF5-A1CC-FAC5A73D9080}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CFA8567B-FA5F-4FF5-A1CC-FAC5A73D9080}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CFA8567B-FA5F-4FF5-A1CC-FAC5A73D9080}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CFA8567B-FA5F-4FF5-A1CC-FAC5A73D9080}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1F36CE5E-843C-4960-8B74-B87DEAB0C730}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1F36CE5E-843C-4960-8B74-B87DEAB0C730}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1F36CE5E-843C-4960-8B74-B87DEAB0C730}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1F36CE5E-843C-4960-8B74-B87DEAB0C730}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6A91F1B2-5C8A-495A-824B-1BC7825B1939}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6A91F1B2-5C8A-495A-824B-1BC7825B1939}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6A91F1B2-5C8A-495A-824B-1BC7825B1939}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6A91F1B2-5C8A-495A-824B-1BC7825B1939}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
